+++ examples/org-unit-tree/index.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+
+import OrgUnitTree from '../../src/org-unit-tree';
+
+
+function OrgUnitTreeExample() {
+    return (
+        <div>
+            <OrgUnitTree />
+        </div>
+    );
+}
+
+ReactDOM.render(<OrgUnitTreeExample />, document.getElementById('app'));