+++ src/org-unit-tree/OrgUnitTree.component.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import log from 'loglevel';
+
+
+class OrgUnitTree extends React.Component {
+    render() {
+        return (
+            <div>
+                Organisation Unit Tree Component coming here Soon&trade;
+            </div>
+        );
+    }
+}
+
+export default OrgUnitTree;