+++ webpack.config.js
@@ -6,6 +6,7 @@ module.exports = {
     entry: {
         'tree-view': './examples/tree-view',
         'data-table': './examples/data-table',
+        'org-unit-tree': './examples/org-unit-tree',
     },
     output: {
         path: path.join(__dirname, '/build'),